move parallel implementation into single file
